# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "API for managing veterinary clinic operations including appointments, medical records, treatments, and more.",
        "title": "Veterinary Management System API",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {
            "name": "API Support",
            "url": "http://www.veterinary-system.io/support",
            "email": "support@veterinary-system.io"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/api/v1",
    "paths": {
...
        "/pets/{id}/surgeries": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Get all surgeries for a specific pet",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Surgeries"
                ],
                "summary": "Get surgeries by pet ID",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Pet ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "List of surgeries for the pet",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/domain.Surgery"
                            }
                        }
                    },
                    "400": {
                        "description": "Invalid pet ID",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
...
        "/surgeries": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Get a list of all surgeries",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Surgeries"
                ],
                "summary": "Get all surgeries",
                "responses": {
                    "200": {
                        "description": "List of surgeries",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/domain.Surgery"
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            },
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Schedule a new surgery for a pet",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Surgeries"
                ],
                "summary": "Create a new surgery",
                "parameters": [
                    {
                        "description": "Surgery details",
                        "name": "surgery",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/domain.Surgery"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Surgery created successfully",
                        "schema": {
                            "$ref": "#/definitions/domain.Surgery"
                        }
                    },
                    "400": {
                        "description": "Invalid request format",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
        "/surgeries/scheduled": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Get all scheduled surgeries within a specified date range",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Surgeries"
                ],
                "summary": "Get scheduled surgeries",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Start date (YYYY-MM-DD)",
                        "name": "start_date",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "End date (YYYY-MM-DD)",
                        "name": "end_date",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "List of scheduled surgeries",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/domain.Surgery"
                            }
                        }
                    },
                    "400": {
                        "description": "Invalid date format or missing parameters",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
        "/surgeries/status": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Get all surgeries filtered by status",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Surgeries"
                ],
                "summary": "Get surgeries by status",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Surgery status (scheduled, in_progress, completed, cancelled)",
                        "name": "status",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "List of surgeries with the specified status",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/domain.Surgery"
                            }
                        }
                    },
                    "400": {
                        "description": "Status parameter is required",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
        "/surgeries/{id}": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Get detailed information about a specific surgery",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Surgeries"
                ],
                "summary": "Get surgery by ID",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Surgery ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Surgery details",
                        "schema": {
                            "$ref": "#/definitions/domain.Surgery"
                        }
                    },
                    "400": {
                        "description": "Invalid surgery ID",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "404": {
                        "description": "Surgery not found",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            },
            "put": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Update information for a specific surgery",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Surgeries"
                ],
                "summary": "Update surgery by ID",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Surgery ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Surgery details",
                        "name": "surgery",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/domain.Surgery"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Surgery updated successfully",
                        "schema": {
                            "$ref": "#/definitions/domain.Surgery"
                        }
                    },
                    "400": {
                        "description": "Invalid request format",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            },
            "delete": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Remove a surgery from the system",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Surgeries"
                ],
                "summary": "Delete surgery by ID",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Surgery ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Surgery deleted successfully",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "400": {
                        "description": "Invalid surgery ID",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
        "/surgeries/{id}/cancel": {
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Mark a surgery as cancelled",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Surgeries"
                ],
                "summary": "Cancel a surgery",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Surgery ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Surgery cancelled successfully",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "400": {
                        "description": "Invalid surgery ID",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "domain.Surgery": {
            "type": "object",
            "properties": {
                "actual_date": {
                    "type": "string"
                },
                "anesthesia_used": {
                    "type": "string"
                },
                "complications": {
                    "type": "string"
                },
                "cost": {
                    "type": "number"
                },
                "created_at": {
                    "type": "string"
                },
                "duration": {
                    "type": "integer"
                },
                "follow_up_date": {
                    "type": "string"
                },
                "follow_up_required": {
                    "type": "boolean"
                },
                "pet": {
                    "description": "Relations",
                    "allOf": [
                        {
                            "$ref": "#/definitions/domain.Pet"
                        }
                    ]
                },
                "pet_id": {
                    "type": "integer"
                },
                "post_op_notes": {
                    "type": "string"
                },
                "pre_op_notes": {
                    "type": "string"
                },
                "scheduled_date": {
                    "type": "string"
                },
                "status": {
                    "$ref": "#/definitions/domain.SurgeryStatus"
                },
                "surgery_id": {
                    "type": "integer"
                },
                "surgery_name": {
                    "type": "string"
                },
                "surgery_type": {
                    "$ref": "#/definitions/domain.SurgeryType"
                },
                "updated_at": {
                    "type": "string"
                },
                "veterinarian": {
                    "$ref": "#/definitions/domain.Veterinarian"
                },
                "veterinarian_id": {
                    "type": "integer"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "BearerAuth": {
            "description": "Type \"Bearer\" followed by a space and JWT token.",
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}